revert delete CNAME
I can see that deleting the CNAME was an attempt to get the
domain to point to Wordpress, but open-austin.org is clearly
still pointing to Github Pages even after the change. I'm reverting
the change because it completely brought down the website, replacing
it with a Github Pages error message.